% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{export}
\alias{export}
\title{Export du fichier final}
\usage{
export(path = ".", fichier = "export_final_GeoNature.xlsx")
}
\arguments{
\item{path}{Le chemin permettant d'accéder aux fichiers}

\item{fichier}{Nom du fichier enregistré}
}
\value{
La synthèse GéoNature finale contenant les colonnes sélectionnées par l'utilisateur
}
\description{
La fonction 'export' permet d’exporter la base de données finale, contenant les variables sélectionnées.
}
\details{
Par défaut, le fichier est appelé "export_final_GeoNature.xlsx" et est stocké au même endroit que le script R.
Il est possible de changer le nom du fichier à l’aide de l’argument 'fichier', en spécifiant l’extension .xlsx (par exemple : fichier.xlsx).
Il est également possible de changer le chemin d’accès à l’aide de l’argument 'path', qui doit être sous la forme «~/Documents» ou «~/Downloads», comme pour la fonction 'import'.
}
