% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{extract}
\alias{extract}
\title{Extraction des colonnes d'intérêt}
\usage{
extract(
  col = c("plante", "caste", "station", "annee_determination", "methode"),
  precision_taxo = NULL
)
}
\arguments{
\item{col}{Colonnes à extraire des champs additionnels. Les colonnes peuvent être : "plante", "caste", "station", "annee_determination" et "methode".}

\item{precision_taxo}{Précision taxonomique souhaitée pour la colonne plante ("famille", "genre" ou "sp").}
}
\value{
La synthèse GéoNature complétée des colonnes sélectionnées
}
\description{
La fonction 'extract' permet d’extraire les colonnes d’intérêt, contenues dans les champs additionnels de la synthèse GéoNature. Les colonnes peuvent être : plante, caste, station, année de détermination et méthode de capture. Un tri des données de plantes pourra également être réalisé dans la fonction.
}
\details{
Marche à suivre :
\itemize{
\item Sélectionner les colonnes à extraire (plante, caste, station, annee_determination, methode) à l’aide de l’argument 'col.' Il est possible de ne sélectionner qu’une colonne ou plusieurs, l’ordre n’ayant pas d’importance.
}

Si une seule colonne est sélectionnée, la fonction s’écrit sous la forme extract(col = "plante").

Dans le cas où plusieurs colonnes sont sélectionnées, la fonction prend la forme extract(col = c("plante", "caste", "annee_determination")) par exemple, c() permettant de combiner des valeurs.

Par défaut, si l’argument "col" n’est pas précisé, toutes les colonnes sont sélectionnées.
\itemize{
\item Sélectionner le degré de précision taxonomique choisie ("famille", "genre" ou "sp") à l’aide de l’argument "precision_taxo". Si un rang taxonomique plus grossier est sélectionné (famille ou genre), les rangs plus précis seront conservés. Par exemple, si on autorise de conserver une précision allant jusqu’à la famille, les observations allant jusqu’au genre ou à l’espèce seront conservées.
}

Par défaut, toutes les observations sont conservées et aucun tri n’est effectué.

La fonction extract peut simplement s’écrire extract(), dans ce cas toutes les colonnes utiles des champs additionnels seront extraites et toutes les données de plantes seront gardées (même les observations NA ou les données ne renvoyant pas à un cd_nom dans TaxRef).

Il faut compter environ 2 minutes d’attente pour l’export de toutes les colonnes et pour une base de données d’environ 100 000 observations
}
